+++ extension.js
@@ -1,3 +1,4 @@
+"use strict";
 /**
  * @todo Prepare for release 1.0.0.
  *   - [ ] Changelog.
@@ -7,30 +8,40 @@
  *   - [ ] Dev docs: semantic versioning.
  *   - [ ] Dev docs: debugging.
  *   - [ ] Dev docs: updating.
+ * @todo Convert to typescript.
+ * @todo Add `deleteNode` option under platformSettings.[platform].
+ * @todo Turn Settings into a class, make channel a private variable, and add loadSettings as a method of the class.
+ * COMPLETE: Replace console.log with channel log.
+ * COMPLETE: Don't import entire library, only what is needed. { workspace }
  */
-// TODO: Don't import entire library, only what is needed. { workspace, OutputChannel }
-const vscode = require("vscode");
+const { workspace, window, commands } = require("vscode");
 const os = require("os");
-let channel = vscode.OutputChannel;
-channel = vscode.window.createOutputChannel("platform-settings");
+let channel = window.createOutputChannel("platform-settings");
 
-// TODO: Turn this into a class, and make channel a private variable.
-// TODO: Replace console.log with channel log.
 //  Inspiration: https://github.com/vscode-restructuredtext/vscode-restructuredtext/blob/460f9f37cdf048e4c30d2705ff9b89ebd03f535b/src/extension.ts
 function loadSettings(platform) {
-  const settingNodes = vscode.workspace.getConfiguration("platformSettings")
-    .platforms[platform].nodes;
-  const config = vscode.workspace.getConfiguration();
+  const settingNodes = workspace.getConfiguration("platformSettings").platforms[
+    platform
+  ].nodes;
+  const removeNodes = workspace.getConfiguration("platformSettings").platforms[
+    platform
+  ].remove;
+  const config = workspace.getConfiguration();
 
-  console.log(`Loading settings for ${platform}.`);
+  channel.appendLine(`Loading settings for ${platform}.`);
   Object.keys(settingNodes).forEach((key) => {
-    console.log(` - ${key}: ${settingNodes[key]}`);
+    channel.appendLine(`  + ${key}: ${settingNodes[key]}`);
     config.update(String(key), settingNodes[key], 1);
+  });
+  Object.keys(removeNodes).forEach((key) => {
+    channel.appendLine(`  - ${key}: ${removeNodes[key]}`);
+    // FIXME: Remove, don't update.
+    config.update(String(key), removeNodes[key], 1);
   });
 }
 
 function updateSettings() {
-  const config = vscode.workspace.getConfiguration("platformSettings");
+  const config = workspace.getConfiguration("platformSettings");
 
   let platform;
   if (os.platform() in config.platforms) platform = os.platform();
@@ -51,21 +62,25 @@ function updateSettings() {
 function activate(context) {
   channel.appendLine("Extension starting...");
 
-  let refreshSettings = vscode.commands.registerCommand(
+  let refreshSettings = commands.registerCommand(
     "extension.refreshSettings",
     function () {
       updateSettings();
     }
   );
 
-  if (vscode.workspace.getConfiguration("platformSettings").autoLoad)
+  if (workspace.getConfiguration("platformSettings").autoLoad) {
+    channel.appendLine(
+      "platformSettings.autoLoad enabled, loading settings..."
+    );
     updateSettings();
+  }
 
   context.subscriptions.push(refreshSettings);
 }
-exports.activate = activate;
 
 function deactivate() {}
+
 module.exports = {
   activate,
   deactivate,